New file src/main.php are the loading steps are executed in a global scope, improve homogeneity inside the class MediaWikiFarm
The global scrope is required for extensions (and skins) using $wgMyConfigParameter = stuff. Nowadays most last versions of
MediaWiki extensions do use the better syntax $GLOBALS['wgMyConfigParameter'] = stuff, or, still better, use the MediaWiki
loading mechanism (wfLoadExtension/wfLoadSkin), but the loading in the global scrope is used here for compatibility with old
extensions or old versions of extensions.

This extension does work as of this version, although some work is still needed to better manage the suffixes and wiki sets
(of the MediaWiki SiteConfiguration context), allowing to view and configure wikis by sets and subsets.