Generalisation of cache files to all files read.
For now it is in the subdirectory '.cache' of the config directory.
Composer is now only loaded when YAML files must be read.

Performance has drastically improved (in ms):

    mean   = 4.7572
    median = 4.3525
    std    = 0.9047
    min    = 3.8898
    max    = 6.9201
    range  = 3.0303
    samples = 50

Performance in the case nothing is cached is about 12ms; it is between
if some files are cached. For information, an equivalent LocalSettings.php
is loaded in 2.9ms (mean). So this extension takes ~1.85ms.

Possibly performance could be still improved by creating a unique file with
all existing wikis (of a farm) with their version associated, although there
is no certainty about that (possibly a big file could take more time than
two small files).