Added the farm in the cache path ; computed performance with a strict protocol: doc added ; removed cache parameter from publicised parameters
Previously there was a collision in cache if two farms used a same file name: fixed.

The cache parameter is no more publicised in default configuration files. The reasoning is:
* the user doesn’t have to know how to configure the cache directory param, that is another
  thing to explain in installation, and a default setting is fine, either developer-fixed,
  either in the main config file
* the cache directory must be known before the main config file is loaded, to cache this file
* it is convenient to have a possibility to disable cache, for developers
* it is convenient to have a possibility to change the cache directory (shared hosts or Windows)
* no cache is not an option
So removing publicity of this core config param is probably the better option: by default there is
a cache in a known location, and if absolutely needed, the location can be changed or the cache
disabled.