Added a mechanism to always use the same version until maintenance/update.php is run
This allows a better consistency: the inventory file can be updated with the new version
(nothing change for now), then update.php is run (nothing change until the end of the script),
and then the new version is used.

This is also a micro-caching mechanism: a small piece of logic is avoided, although this
only minor.