Renamed entry points and make them inactive when directly called from their normal directory
These entry points are only useful in the case of a multiversion installation; so
as a classical MediaWiki extension, they must be disabled to avoid any security issue.

Updated doc installation accordingly

The directory /www is still git-ignored but the entry points inside are forced to be
version-controled. This allows to version-control independently the directory /www, e.g.
for 404 error pages.

Change-Id: I7d712efa2b1bfc5ae61a342b61aeb66ccda797db