Version 0.2.0
Reorganised the main class to make it testable:
* almost all methods become public;
* all object properties become private but readable;
* added 2x15 tests, in mono- and multiversion modes.

Object properties are more rational: no more duplicate between $params and $variables,
$params become raw unchanged config (its name become $farmConfig) and $variables become
its request-specific alter-ego. The subarray $params['globals'] (for MW configuration)
become an independant property $configuration.

Removed the object property $emergency; added exceptions, either:
* LogicException (and children) for code-issue exceptions;
* dedicated MWFConfigurationException (child of RuntimeException) for configuration-level
  issues (either in farm configuration or server-PHP issue).
All these exceptions result in a HTTP 500 error. Softer issues like a missing wiki
still result in a HTTP 404 error.

A control flow graph was created (on paper, I still have to draw it on Inkscape) to check
there was no loop: there is no. This resulted also in a rewritting of selectFarm to make it
standalone, and a rewritting of checkExistence. I tagged most of the methods with two custom
PHPDoc tags: mediawikifarm-const and mediawikifarm-idempotent (definitions on top of file)
to better see where the data are changed.

2x15 tests are added, corresponding to mono- and multiversion installations. For now it only
tests the first part of the extension (construction and selection+existence of the wiki); the
second part (MW configuration) still has to be written (but I would like to write dedicated
code for this instead of using wgConf, mainly to optimise the 'compilation' and to become
compatible with MW < 1.6).

Fixed a bug when 'deployments' key is in config but we are in monoversion mode.
Fixed some edge-edge cases with the aid of unit tests.
Removed the use of DIRECTORY_SEPARATOR (ugly; open a bug if Windows growls).
Added a function isMediaWiki checking if a directory contains MW (with rationale in doc).

A next commit will follow to sort methods and some files -- I avoid to do it with these
big changes to have readable diffs.

Change-Id: I5411ef6fbe17205b9415f08e715ac05166758c5f