Whitelist the SAPI PHPDBG for PHPUnit with PHP 7
PHPUnit, when used with PHP 7, can use either the CLI or PHPDBG SAPIs to
generate code coverage. PHPDBG gives good and almost similar results
compared to CLI+Xdebug, and it is more pratical in a dev environment
not to enable Xdebug on CLI for performance reasons.

BTW added a default argument to MediaWikiFarm::getVariable() and used it.

Change-Id: I80fd8c54c1ed4f5d6b6032e2b60ccb7dfd91989c