Added unit tests and improved code coverage; fixed edge cases
The number of tests increases from 36 to 81, assertions from 64 to 121.
The code coverage (with strict code coverage activated) increased from 21% to
33% in number of lines and 27% to 42% in number of functions/methods; the CRAP
decreased from 17k to 8k. Some remaining methods of the first part of the
extension (loading and version branching) have to be tested, but the big remaining
part is the second part of the extension (compilation and loading of the config);
a config file with the main parameter types is already prepared.

Note there are a few global functions (3); PHP_CodeCoverage didn’t count them in
the number of functions+methods; the stats above are obtained from a patched version
of PHP_CodeCoverage (see [1]).

Some edge cases have been fixed with the aid of unit tests, particularly around
file reading and error handling.

Removed the dependency on the global variable $mwfScript used in mwscript.php by
adding a new fixed property $entryPoint (it does not change after construction).

Removed caching using serialisation (only caching with PHP files remains); was
dead unused and probably unuseful code.

[1] https://github.com/sebastianbergmann/php-code-coverage/issues/468

Change-Id: I137bacca746dbeae28e023e4bbd49ebd78abb054