Implement a dedicated engine to compile configuration
Not activated for now, but running.

This is similar in idea to wgConf, but oriented to manage multiple
config files with different prefixes, with priorities walking accross
config files.

This also split out '+'-prefixed arrays to enable cross-compilation
of config (when the compiling MW is not the target MW).

This will also enable MediaWikiFarm to work with MediaWiki 1.6-
when wgConf did not exist.

Change-Id: I02e29ba742dfd049d988d912114f26a0485944d6