Added a mechanism to cache server name in the first step (server recognition)
The second step was already cached (a per-wiki LocalSettings.php is created
according to its origin files), but the first step (server recognition) was not.
This is now the case: when a wiki is requested, its core variables (identifier,
version, code path, etc.) are computed and stored in a "cache file"; next times,
the host is recognised and its data are loaded. Note the origin files can be
different from a host to another due to the dynamic hierarchical loading, so these
origin files are themselves stored and validation is done after the loading
(if it is a cache miss, we have loaded these data for nothing but it is however
better to check cache freshness afterwards than never).

Also:
* added a JSON schema for the array stored in this new cache file
* added a parameter in MWF::readFile to avoid cache a file after reading
  (needed for this new feature)
* origin files used to create a decision for the existence of a certain host
  are registered in order to put them in the cache file and self-check the
  cache freshness ('coreconfig' in the code)
* added a system variable $FARM storing the farm name (needed for this new
  feature)

Schema change:
* renamed the property 'exec' to 'executable', more understandable

Bugs:
* fixed a bug in PHP 5.6 leading to a wrong extension loading mechanism due to
  a race condition: in these case we explicitely indicate the loading mechanism
  (introduced some days ago) with a parameter wgUseMyExtension, another mechanism
  could be used, because in a foreach loop where new keys are added, PHP 5 also
  evaluate these new keys (at the contrary of PHP 7 probably)
* AbstractMediaWikiFarmScript::copyr was not static although it should

Change-Id: I20f56c034975ed1c39da99dfc2f4f5c8bac2a8b8