Add a logging infrastructure
Some errors and exceptions were already collected but not logged; now all these errors
are logged with syslog with a syslog tag to be defined (by default logging is
activated with the tag 'mediawikifarm').

The logging code is splitted in two methods for testing purposes since the second is
harly testable. I chose not to use PSR-3 infrastructure to keep code as small as possible
and avoid dependencies; and syslog messages can be redirected to a specific destination
with appropriate system syslog configuration.

Improved some messages with some additionnal informations and added two messages.

“Errors” (e.g. unreadable files) are registered in the same internal variable, but possibly
a specific internal variable will be created in the future.

Change-Id: I30601fba0eb9ba11433f59a58b947a57c59b581b