Fix bug about empty lists/objects in composer.json
During installation of Composer-managed MediaWiki extensions, the composer.json
is changed and sometimes the 'require' section becomes empty; this was
previously solved by adding the option JSON_FORCE_OBJECT in json_encode, but
this option transforms all lists to objects, which conflicts with standard
MediaWiki composer.json which contains real listes (e.g. 'authors') and Composer
stops because of schema error.

The solution implemented here is to remove the option JSON_FORCE_OBJECT and
remove the sections 'require-dev' and 'require' when they are empty to avoid
they are changed to an empty list. Added a list in the test composer.json to
avoid further regressions about this.

Change-Id: Id4b4017f69fd611b8b0e84b7b0b27ce3d326e624