Fix bug when a mwcomposer was used but no Composer-managed extension is activated
When mwcomposer was used (hence a special vendor/autoload.php is installed) and
when no Composer-managed extension is activated, no Composer libraries is loaded
and MediaWiki is obviously complaining because there is no PSR-3 library.

The fix is simply to check if there are Composer-managed extensions activated
and else use the minimal Composer autoloader.

Change-Id: I3cf5e2826a92c0f27226621ca0259901299c65e3