Execute mwcomposer just like composer
Previously mwcomposer selected the directory according to a configured wiki,
but it introduces a chicken-and-egg problem when you want to upgrade: you
create a new MediaWiki installation in another directory, change some wiki
to this new version even if no Composer was run (and MediaWiki publicly
complains on the Web), you run mwcomposer with this wiki, and then it works,
and you can run update.php for each wiki.

The trick could be to use a junk wiki, but this is now solved: now you
prepare your new MediaWiki version, run mwcomposer just like you would have
run composer, and then switch each wiki to this new version and run update.php
for each wiki.

By the way, wrote some documentation in docs/scripts.rst.

Change-Id: I150e805bc32f7329784a501e75d7faf218e389e4