Refactoring: extraction of configuration compilation
The main class MediaWikiFarm had a lot of responsabilities, one of those was
to compile configuration for a specific wiki. This commits splits out this
configuration compilation into a dedicated class. In number of lines, the
main class MediaWikiFarm drops from 2030 to 1427.

This is a first step to more refactoring to create smaller functions and
possibly multiple classes, in order to facilitate long-term maintenance and
improve understanding. Possibly the 'existence' feature inside the main
class will be also splitted out into a dedicated class, which would become
sort of a (dependency injection) container.

BTW, updated PHPCS rule exceptions for PHPCS-MediaWiki 0.8.

Change-Id: I54d4d2739215f9cc4f59fb2671d2cd7e797d16f9