Define more precisely the set of existing wikis
Given there are two mechanism to check if a wiki exists (“variables” files and
“versions” file) and given the two mechanisms answer “possibly existing” if a
wiki is not registered in their mechanism, it was possible to define ghost
wikis, registered nowhere but with some default coarse-defined parameters (MW
parameters like a default database for a suffix, catching there all ghost
wikis).

This commit fixes this odd behaviour: wikis must be defined in at least one of
the two mechanisms (the first mechanism has precedence). More precisely either
the 'wikiID' is defined with at least one identifying variable checked against
a file of existing values, either the ”versions“ file must contain the wikiID.

BTW, remove the need to define the “variables” mechanism (it is possible to
entirely rely on the “versions” mechanism); it will be a bit more easier to
setup a basic installation as described in the “Quick start” guide.

Beyond this improvement in the definition, this is required to create
automatic lists of wikis; without it, the ghost wikis issue remains.

Also, strenghten some PHPCS exceptions and add some gitignored files
temporarily used in tests.

Change-Id: I569c780365022632d5db32e5406fbed86756290c