Split out utilitary functions into dedicated classes
This reduces the complexity of the main class (CRAP from 253 to 202) and make
possible the use of these functions without creating the main object. In
particular I need the function readFile for the “list wikis” feature, but
don’t need the main object.

By the way, the function dedicated to read YAML files is enclosed in a class,
so that it is a general design to have utils functions in classes.

Small thing: I remove the @author given there will be many authors in the
long term (and Git history lists the authors), and it removes one publication
on the Web of my email address, hence a source of information for spammers.

Change-Id: Id9a47bf64061ab6d292e338c750d311f96be0e32