Cache files in /cache/wikis were never used
Due to the change in T161413, cache file in /wikis were differentiated
according to the path, but only the prefix of the path should be used, which
was not the case when reading the cache file.

To fix this and go further, a new cache file is created /cache/wikis.php,
which contains the list of known hosts and paths. The paths are a regex
instead of an array for an easier implementation. This new file is also an
additional security check: the existence of a wiki is checked
programmatically instead of a file system check.

Bug: T175332
Change-Id: I53e044ddc163cd614fff198e5ee12f795660bedb