Always check Composer extensions independantly of ExtensionRegistration
Sometimes, when MediaWikiFarm is in monoversion mode, ExtensionRegistry
is known, and we should check if some extension should be loaded with
Composer; it was no more the case since ca185e0.

Also, the function MediaWikiFarmConfiguration::detectLoadingMechanism
no more tries to detect Composer extension given the case is handled
in an other way.

Change-Id: I6b1379826f698861d66016c836b3d867d2d85fec