Enumerate all known wikis
A CLI script is added to output the list of wikis known by MediaWikiFarm
in its configuration files. This is achieved by reading the server
regexes, reading the values of the variables, and re-creating the URLs.
Thanks to the previous commit c8124a3 all wikis are always explicitely
known, hence this method will create exhaustive lists, except one
mechanism is not currently implemented in this feature (the “versions”
mechanism).

A class MediaWikiFarmList is created for this feature; this is the “model”
class for this feature.

It should be added in the future an extension in mwscript.php to run a
MediaWiki script on all or part of a farm; a difficulty to take into account
is that independant scripts must be launched (not in the current PHP script)
because wikis are possibly in different MediaWiki versions and it could
create “redefinition of classes” errors.

Bug: T162737
Change-Id: I48a0898160d6aab0338c29d7e9ddca24c81d81df