Stop MediaWikiTestCase defition appearing for IDEs
This has been annoying me for a while now.
As I have core and all extensions checked out, my IDE
detects this defition of MediaWikiTestCase as the first,
and as a result all of my typhinting for the actual class
is off.

This patch uses a class alias instead of defining a class and
wraps that in a call_user_func to hide it from IDEs

Change-Id: Iadb86b051da707fa6d6ecee8b7ffe38bd87fff63