Execute a script accross multiple wikis
This can be executed with:
  php bin/mwscript.php --wiki=* lag

Also, a subset of wikis can be chosen:
  php bin/mwscript.php --wiki=*.example.org lag

Each execution for a specific wiki is a dedicated run; hence
multiple MW versions can coexist and a failing run has no
consequences to other runs. Each line in the output is prefixed
by the wiki URL, and the returned value is 0 iff all returned values
(for each wiki) is 0.

Also, added an option -q (quiet) for individual runs to disable
the header identifying the wiki and version. This option is
automatically enabled for execution on multiple wikis.

Bug: T162737
Change-Id: I832dff4ad6c2aeb34a7967d98ea0d31416699b7b