Version 1.0.0
(reissuing since version must be written on 2 locations.)

Although this extension was stable since a long time, it was not
administratively declared as 1.0.0. Now it is.

Well, mainly because a breaking change should occur to support only
PHP 7.0+, so the new one will be 2.0.0, but versions ^1.0.0 will be
still supported.

Version ^1.0.0 supports:
* PHP 5.2+
* MediaWiki 1.3+

Change-Id: Iea6c0e3b263b8e29e8fd889ebedf8069cecb329b